// src/app/core/api/model/policyDestinationInputDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PolicyDestinationInputDto { 
    destinationId: string;
    role: PolicyDestinationInputDto.RoleEnum;
    priority?: number;
    retentionDaysOverride?: number;
    retentionMaxCopiesOverride?: number;
}
export namespace PolicyDestinationInputDto {
    export const RoleEnum = {
        Primary: 'primary',
        Replica: 'replica'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}